Make authorization lead into form
