+++ src/views/apply/form.html
@@ -1,8 +1,8 @@
 {% extends 'default.html' %}
 {% block pageContent %}
-Thanks! We just need a few more details about you
 <form class="apply-form" method="POST" enctype="multipart/form-data">
   <div class="form-content">
+    Hello, {{ user.first_name }}! Let us get to know you better
     {{ formHtml|safe }}
   </div>
   <button type="submit">Submit</button>
