+++ src/views/apply/index.html
@@ -2,7 +2,7 @@
 {% set title = 'Apply | Hack Cambridge' %}
 {% block pageContent %}
 <h1>Apply to Hack Cambridge</h1>
-<a href="/apply/mymlh">
+<a href="/apply/auth">
   <button>
     Log in with MyMLH
   </button>
